refactor: page views and likes are blog components
